+++ types/core/converter/ExportFileAeroflyCustomMissionsTmcConverter.d.ts
@@ -1,4 +1,3 @@
-import type { AeroflyAircraft } from "@fboes/aerofly-data/data/aircraft-liveries.json";
 import { AeroflyFlight } from "@fboes/aerofly-custom-missions";
 import { ExportFileConverter } from "./ExportFileConverter.js";
 import { AeroflyNavRouteBase } from "@fboes/aerofly-custom-missions/types/dto-flight/AeroflyNavRouteBase.js";
@@ -6,6 +5,5 @@ export declare class ExportFileAeroflyCustomMissionsTmcConverter implements Expo
     static readonly fileExtension = "tmc";
     convert(flightplan: AeroflyFlight): string;
     getWaypointType(w: AeroflyNavRouteBase): import("@fboes/aerofly-custom-missions/types/dto/AeroflyMissionCheckpoint.js").AeroflyMissionCheckpointType;
-    protected findAircraftData(aeroflyCodeAircraft: string): AeroflyAircraft | undefined;
 }
 //# sourceMappingURL=ExportFileAeroflyCustomMissionsTmcConverter.d.ts.map