+++ types/core/formatter/AeroflyFlightFormatter.d.ts
@@ -1,8 +1,17 @@
 import { AeroflyFlight } from "@fboes/aerofly-custom-missions";
-import type { AeroflyAircraft, AeroflyAircraftLivery } from "@fboes/aerofly-data/data/aircraft-liveries.json";
 export declare class AeroflyFlightFormatter {
-    static getAircraft(currentAircraft: AeroflyAircraft | undefined, currentLivery: AeroflyAircraftLivery | undefined): string;
+    static getAircraft(aeroflyFlight: AeroflyFlight): string;
     static getFuelAndPayload(aeroflyFlight: AeroflyFlight): string;
+    static getFlightplanIdentifier(aeroflyFlight: AeroflyFlight): string;
+    static getFlightplanOriginCode(aeroflyFlight: AeroflyFlight): string;
+    static getFlightplanDestinationCode(aeroflyFlight: AeroflyFlight): string;
+    static getFlightplanSummary(aeroflyFlight: AeroflyFlight): string;
+    static getFlightplanWaypoints(aeroflyFlight: AeroflyFlight): string;
+    static getFlightplanDistance(aeroflyFlight: AeroflyFlight): string;
+    static getWind(aeroflyFlight: AeroflyFlight): string;
+    static getTemperature(aeroflyFlight: AeroflyFlight): string;
+    static getVisibility(aeroflyFlight: AeroflyFlight): string;
+    static getClouds(aeroflyFlight: AeroflyFlight): string;
     static numberToString(num: number): string;
     static dateToString(date: Date): string;
 }