+++ types/core/util/AeroflyFlightHelper.d.ts
@@ -1,11 +1,13 @@
 import { AeroflyFlight } from "@fboes/aerofly-custom-missions";
 export declare class AeroflyFlightHelper {
-    protected aeroflyFlight: AeroflyFlight;
-    constructor(aeroflyFlight: AeroflyFlight);
     /**
      *
      * @returns in meters
      */
-    getFlightplanDistance(): number;
+    static getFlightplanDistance(aeroflyFlight: AeroflyFlight): number;
+    /**
+     * @returns nautical time zone offset based on the coordinates of the departure airport
+     */
+    static getDepartureTimeZone(aeroflyFlight: AeroflyFlight): number;
 }
 //# sourceMappingURL=AeroflyFlightHelper.d.ts.map